refactor: Break down evaluator.rs into smaller, focused modules
Split the large evaluator.rs file (1586 lines) into a modular structure
for improved maintainability and code organization.

**Module Structure:**
- `evaluator/core.rs` (~200 lines): Struct definitions, constructors, and binary operations
- `evaluator/expressions.rs` (~350 lines): ExpressionEvaluator::eval() implementation
- `evaluator/combined.rs` (~350 lines): CombinedExpressionEvaluator::eval() for JOINs
- `evaluator/casting.rs` (~250 lines): Type conversion and CAST operations
- `evaluator/pattern.rs` (~60 lines): LIKE pattern matching
- `evaluator/tests.rs` (~130 lines): Unit tests
- `evaluator/mod.rs`: Module declarations and public API re-exports

**Benefits:**
- Better separation of concerns - each module has a single responsibility
- Easier to navigate and understand specific functionality
- Reduced merge conflicts when multiple developers work on different features
- More testable code with clear module boundaries
- Maintains 100% API compatibility - all existing tests pass

**Testing:**
✅ All 22 end-to-end tests pass
✅ Build completes successfully
✅ No behavioral changes - pure refactoring

Closes #158
